Descriptor structures and extending renderers
 - Dynamically generated module header
 - Addition of utils and vulkan public docs groups
 - Replacement of th_Status_t struct with int returns
 - API-dependent extension options & extension descriptor type
 - Descriptor structures (e.g. th_RendererDescriptor_t)
 - API-agnostic version structure
 - Vulkan instance extensions and layers support and validation
 - Vulkan render system
 - Vulkan instance creation (contained within render system)

+++ lib/state/state.h
@@ -17,15 +17,13 @@
     extern "C" {
 #endif // __cplusplus
 
-#include <stdint.h>
-
 /**
  * @ingroup state
  * @brief Internal library state structure.
  */
 typedef struct th_State_t {
     /// @brief th_DebugSeverity_t bit-field holding severities of messages to display.
-    uint8_t debugSeverityFilter;
+    int debugSeverityFilter;
 } th_State_t;
 
 /**
